meson: Install Tracker-1.0 GIR and Typelib
This is done in a horrendously hacky way for the time being. We really
need accessor functions for target outputs to do it more cleanly, so
we could call .vala_gir() on the libtracker_sparql target object and
get the actual GIR output.

See https://github.com/mesonbuild/meson/issues/891